tuiをファイル分割
