+++ tui/status.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <string>
+
+inline std::string &defaultStatus() {
+    static std::string s = "↓/↑ or J/K = move, Ctrl+C = quit";
+    return s;
+}
