+++ tui/text.h
@@ -1,33 +1,16 @@
 #pragma once
+#include <ftxui/dom/elements.hpp>
+#include <ftxui/component/component.hpp>
+#include <ftxui/component/screen_interactive.hpp>
+#include <webcface/text.h>
+#include <string>
+#include <memory>
 
-inline ftxui::Component textComponent(const webcface::Text &text,
-                                      std::shared_ptr<std::string> help,
-                                      std::shared_ptr<ftxui::Element> result,
-                                      bool /*light*/) {
-    return ftxui::Renderer([=](bool focused) {
-        if (focused) {
-            *help = defaultStatus();
-        }
-        return ftxui::hbox({
-                   ftxui::text(text.member().name()),
-                   ftxui::text(":"),
-                   ftxui::text(text.name()),
-                   ftxui::text(" = "),
-                   ftxui::text(text.get()) |
-                       (focused ? ftxui::bold : ftxui::nothing) | ftxui::xflex,
-               }) |
-               ftxui::xflex | (focused ? ftxui::focus : ftxui::nothing);
-    });
-}
-inline void
-addTextComponent(ftxui::ScreenInteractive &screen, ftxui::Component &container,
-                 const webcface::Text &text, std::shared_ptr<std::string> help,
-                 std::shared_ptr<ftxui::Element> result, bool light) {
-    auto handle = std::make_shared<webcface::Text::EventHandle>();
-    *handle = text.prependListener([=, &screen, &container] {
-        screen.Post([=, &container] {
-            container->Add(textComponent(text, help, result, light));
-        });
-        text.removeListener(*handle);
-    });
-}
+ftxui::Component textComponent(const webcface::Text &text,
+                               std::shared_ptr<std::string> help,
+                               std::shared_ptr<ftxui::Element> result,
+                               bool light);
+void addTextComponent(ftxui::ScreenInteractive &screen,
+                      ftxui::Component &container, const webcface::Text &text,
+                      std::shared_ptr<std::string> help,
+                      std::shared_ptr<ftxui::Element> result, bool light);
