+++ tui/view.h
@@ -1,7 +1,12 @@
 #pragma once
-#include "color.h"
-#include "func.h"
-#include <cctype>
+#include <ftxui/dom/elements.hpp>
+#include <ftxui/component/component.hpp>
+#include <ftxui/component/screen_interactive.hpp>
+#include <webcface/view.h>
+#include <string>
+#include <memory>
+#include <vector>
+#include <unordered_map>
 
 class ViewUIContainer : public ftxui::ComponentBase {
     webcface::View view;
@@ -16,285 +21,18 @@ class ViewUIContainer : public ftxui::ComponentBase {
     // childrenにレイアウトを反映 (renderはしない)
     // childrenは2次元vectorのような感じになる
     // 要素が足りなければAddし、多ければDetachする
-    void updateLayout() {
-        auto &root = this->ChildAt(0);
-        // childrenを空にする (行は削除しない)
-        for (std::size_t row = 0; row < root->ChildCount(); row++) {
-            root->ChildAt(row)->DetachAllChildren();
-            has_no_component[row] = false;
-        }
-        if (root->ChildCount() == 0) {
-            root->Add(ftxui::Container::Horizontal({}));
-            has_no_component.push_back(false);
-        }
-        std::size_t row = 0;
-        for (const auto &cp : this->view.get()) {
-            switch (cp.type()) {
-            case webcface::ViewComponentType::text:
-                break;
-            case webcface::ViewComponentType::new_line:
-                // 空の行もfocusを取れるようにする(仮)
-                if (root->ChildAt(row)->ChildCount() == 0) {
-                    has_no_component[row] = true;
-                    root->ChildAt(row)->Add(
-                        ftxui::Renderer([](bool /*focused*/) {
-                            return ftxui::emptyElement();
-                        }));
-                }
-                if (root->ChildAt(row)->ChildCount() > 0) {
-                    row++;
-                }
-                if (root->ChildCount() <= row) {
-                    root->Add(ftxui::Container::Horizontal({}));
-                    has_no_component.push_back(false);
-                }
-                break;
-            case webcface::ViewComponentType::button: {
-                ftxui::Component ui_cp = this->ui_components[cp.id()];
-                if (!ui_cp && cp != this->prev_components[cp.id()]) {
-                    this->prev_components[cp.id()] = cp;
-                    auto option = ftxui::ButtonOption::Animated(
-                        convertColor(cp.bgColor(), webcface::ViewColor::green,
-                                     light),
-                        convertColor(cp.textColor(), webcface::ViewColor::black,
-                                     light),
-                        convertColor(cp.bgColor(), webcface::ViewColor::green,
-                                     light, true),
-                        convertColor(cp.textColor(), webcface::ViewColor::black,
-                                     light));
-                    option.transform = [](const ftxui::EntryState &s) {
-                        if (s.focused) {
-                            return ftxui::text("(" + s.label + ")") |
-                                   ftxui::bold;
-                        } else {
-                            return ftxui::text("(" + s.label + ")");
-                        }
-                    };
-                    this->ui_components[cp.id()] = ui_cp = ftxui::Button(
-                        cp.text(),
-                        [func = cp.onClick(), result = this->result] {
-                            if (func) {
-                                runAsync(*func, result);
-                            }
-                        },
-                        option);
-                }
-                root->ChildAt(row)->Add(ui_cp);
-                break;
-            }
-            case webcface::ViewComponentType::text_input:
-            case webcface::ViewComponentType::decimal_input:
-            case webcface::ViewComponentType::number_input: {
-                ftxui::Component ui_cp = this->ui_components[cp.id()];
-                if (!ui_cp && cp != this->prev_components[cp.id()]) {
-                    this->prev_components[cp.id()] = cp;
-                    auto bind_ref = std::make_shared<std::string>();
-                    if (cp.bind()) {
-                        *bind_ref = cp.bind()->get();
-                        cp.bind()->appendListener(
-                            [bind_ref](const auto &b) { *bind_ref = b; });
-                    }
-                    auto content_ref = std::make_shared<std::string>();
-                    auto is_error = std::make_shared<bool>(false);
-                    ftxui::InputOption option{};
-                    option.content = ftxui::StringRef(content_ref.get());
-                    option.multiline = false;
-                    option.on_change = [cp, is_error, content_ref] {
-                        *is_error = true;
-                        switch (cp.type()) {
-                        case webcface::ViewComponentType::text_input:
-                            if ((cp.min() && content_ref->size() < cp.min()) ||
-                                (cp.max() && content_ref->size() > cp.max())) {
-                                return;
-                            }
-                            break;
-                        case webcface::ViewComponentType::decimal_input: {
-                            double val = 0;
-                            for (char c : *content_ref) {
-                                if (!std::isdigit(c) && c != '.') {
-                                    return;
-                                }
-                            }
-                            try {
-                                val = std::stod(*content_ref);
-                            } catch (const std::exception &e) {
-                                return;
-                            }
-                            if ((cp.min() && val < cp.min()) ||
-                                (cp.max() && val > cp.max())) {
-                                return;
-                            }
-                            break;
-                        }
-                        case webcface::ViewComponentType::number_input: {
-                            int val = 0;
-                            for (char c : *content_ref) {
-                                if (!std::isdigit(c)) {
-                                    return;
-                                }
-                            }
-                            try {
-                                val = std::stoi(*content_ref);
-                            } catch (const std::exception &e) {
-                                return;
-                            }
-                            if ((cp.min() && val < cp.min()) ||
-                                (cp.max() && val > cp.max())) {
-                                return;
-                            }
-                            break;
-                        }
-                        default:
-                            break;
-                        }
-                        *is_error = false;
-                    };
-                    if (cp.onChange()) {
-                        option.on_enter = [content_ref, func = *cp.onChange(),
-                                           is_error, result = this->result] {
-                            if (!*is_error) {
-                                runAsync(func, result, *content_ref);
-                            }
-                        };
-                    }
-                    option.transform = [content_ref, bind_ref, is_error,
-                                        light = this->light](
-                                           ftxui::InputState state) {
-                        state.element = ftxui::hbox({
-                            ftxui::text("["),
-                            state.element |
-                                (state.focused ? ftxui::bold : ftxui::nothing) |
-                                ftxui::xflex,
-                            ftxui::text("]"),
-                        });
-                        state.element |= ftxui::color(
-                            *is_error
-                                ? convertColor(webcface::ViewColor::red,
-                                               webcface::ViewColor::red, light)
-                                : convertColor(webcface::ViewColor::black,
-                                               webcface::ViewColor::black,
-                                               light));
-                        if (state.focused || state.hovered) {
-                            state.element |= ftxui::bgcolor(
-                                convertColor(webcface::ViewColor::white,
-                                             webcface::ViewColor::white, light,
-                                             state.hovered));
-                        }
-                        if (!state.focused) {
-                            *content_ref = *bind_ref;
-                        }
-                        return state.element;
-                    };
-                    auto ui_input = ftxui::Input(option);
-                    this->ui_components[cp.id()] = ui_cp =
-                        ftxui::Renderer(ui_input, [ui_input] {
-                            return ui_input->Render() |
-                                   ftxui::size(ftxui::WIDTH, ftxui::EQUAL, 10);
-                        });
-                }
-                root->ChildAt(row)->Add(ui_cp);
-                break;
-            }
-            case webcface::ViewComponentType::select_input: {
-                ftxui::Component ui_cp = this->ui_components[cp.id()];
-                if (!ui_cp && cp != this->prev_components[cp.id()]) {
-                    this->prev_components[cp.id()] = cp;
-                    auto bind_ref = std::make_shared<std::string>();
-                    if (cp.bind()) {
-                        *bind_ref = cp.bind()->get();
-                        cp.bind()->appendListener(
-                            [bind_ref](const auto &b) { *bind_ref = b; });
-                    }
-                    ftxui::DropdownOption option{};
-                    auto options = std::make_shared<std::vector<std::string>>();
-                    for (const auto &o : cp.option()) {
-                        options->push_back(o.asStringRef());
-                    }
-                    option.radiobox.entries = options.get();
-                    auto selected = std::make_shared<int>();
-                    option.radiobox.selected = selected.get();
-                    if (cp.onChange()) {
-                        option.radiobox.on_change =
-                            [cp, func = *cp.onChange(), options_v = cp.option(),
-                             bind_ref, selected, result = this->result] {
-                                auto selected_v = options_v.at(*selected);
-                                if (*bind_ref != selected_v) {
-                                    runAsync(func, result, selected_v);
-                                }
-                            };
-                    }
-                    option.checkbox.transform =
-                        [bind_ref, options_v = cp.option(), options,
-                         selected](const ftxui::EntryState &state) {
-                            auto t = ftxui::text(state.label);
-                            if (!state.focused) {
-                                for (std::size_t i = 0; i < options_v.size();
-                                     i++) {
-                                    if (options_v[i] == *bind_ref) {
-                                        *selected = i;
-                                    }
-                                }
-                            }
-                            if (state.focused) {
-                                t |= ftxui::bold;
-                            }
-                            return ftxui::hbox(
-                                {ftxui::text(state.state ? "[↓ " : "[→ "), t,
-                                 ftxui::text("]")});
-                        };
-                    option.radiobox.transform = [](const ftxui::EntryState &s) {
-                        auto prefix = ftxui::text(s.state ? "* " : "  ");
-                        auto t = ftxui::text(s.label);
-                        if (s.focused) {
-                            t |= ftxui::bold;
-                        }
-                        return ftxui::hbox({prefix, t});
-                    };
-                    option.transform = [options, light = this->light](
-                                           bool open, ftxui::Element checkbox,
-                                           ftxui::Element radiobox) {
-                        if (open) {
-                            return ftxui::vbox({
-                                checkbox,
-                                radiobox | ftxui::vscroll_indicator |
-                                    ftxui::yframe |
-                                    ftxui::size(ftxui::HEIGHT, ftxui::LESS_THAN,
-                                                5) |
-                                    ftxui::bgcolor(
-                                        convertColor(webcface::ViewColor::white,
-                                                     webcface::ViewColor::white,
-                                                     light, true)),
-                                ftxui::filler(),
-                            });
-                        }
-                        return ftxui::vbox({checkbox, ftxui::filler()});
-                    };
-                    auto ui_dropdown = ftxui::Dropdown(option);
-                    this->ui_components[cp.id()] = ui_cp = ui_dropdown;
-                }
-                root->ChildAt(row)->Add(ui_cp);
-                break;
-            }
-            }
-        }
-        if (root->ChildAt(row)->ChildCount() == 0) {
-            has_no_component[row] = true;
-            root->ChildAt(row)->Add(ftxui::Renderer(
-                [](bool /*focused*/) { return ftxui::emptyElement(); }));
-        }
-    }
+    void updateLayout();
+
+    ftxui::Component buttonComponent(const webcface::ViewComponent &cp) const;
+    ftxui::Component inputComponent(const webcface::ViewComponent &cp) const;
+    ftxui::Component dropdownComponent(const webcface::ViewComponent &cp) const;
 
   public:
     explicit ViewUIContainer(const webcface::View &view,
                              std::shared_ptr<std::string> help,
-                             std::shared_ptr<ftxui::Element> result, bool light)
-        : ftxui::ComponentBase(), view(view), ui_components(),
-          prev_components(), focused_row(0), has_no_component(), help(help),
-          result(result), light(light) {
-        this->Add(ftxui::Container::Vertical({}, &focused_row));
-        view.prependListener([this] { this->updateLayout(); });
-        this->updateLayout();
-    }
+                             std::shared_ptr<ftxui::Element> result,
+                             bool light);
+
     ~ViewUIContainer() override {}
     ViewUIContainer(const ViewUIContainer &) = delete;
     ViewUIContainer(ViewUIContainer &&) = delete;
@@ -302,100 +40,16 @@ class ViewUIContainer : public ftxui::ComponentBase {
     ViewUIContainer &operator=(ViewUIContainer &&) = delete;
 
     ftxui::Element RenderCol(const std::vector<ftxui::Element> &elements_col,
-                             std::size_t current_row) const {
-        if (this->Focused() && focused_row == current_row &&
-            has_no_component[current_row]) {
-            return ftxui::flexbox(elements_col) | ftxui::bold | ftxui::focus;
-            // ftxui::bgcolor(ftxui::Color::GrayDark)
-        } else {
-            return ftxui::flexbox(elements_col);
-        }
-    }
+                             std::size_t current_row) const;
     // renderする (childrenのレイアウトは無視)
-    ftxui::Element Render() override {
-        if (this->Focused()) {
-            *help = defaultStatus();
-        }
-        std::vector<ftxui::Element> elements, elements_col;
-        for (const auto &cp : this->view.get()) {
-            switch (cp.type()) {
-            case webcface::ViewComponentType::text:
-                elements_col.push_back(ftxui::text(cp.text()));
-                break;
-            case webcface::ViewComponentType::new_line:
-                elements.push_back(RenderCol(elements_col, elements.size()));
-                elements_col.clear();
-                break;
-            case webcface::ViewComponentType::button:
-            case webcface::ViewComponentType::text_input:
-            case webcface::ViewComponentType::decimal_input:
-            case webcface::ViewComponentType::number_input:
-            case webcface::ViewComponentType::select_input: {
-                ftxui::Component ui_cp = this->ui_components[cp.id()];
-                if (ui_cp) {
-                    elements_col.push_back(ui_cp->Render());
-                    if (ui_cp->Focused()) {
-                        std::stringstream ss;
-                        switch (cp.type()) {
-                        case webcface::ViewComponentType::button:
-                            ss << "Enter = click button";
-                            break;
-                        case webcface::ViewComponentType::text_input:
-                            if (cp.min()) {
-                                ss << "min. " << *cp.min() << " chars, ";
-                            }
-                            if (cp.max()) {
-                                ss << "max. " << *cp.max() << " chars, ";
-                            }
-                            ss << "Press Enter to send";
-                            break;
-                        case webcface::ViewComponentType::decimal_input:
-                        case webcface::ViewComponentType::number_input:
-                            if (cp.min()) {
-                                ss << "min. " << *cp.min() << ", ";
-                            }
-                            if (cp.max()) {
-                                ss << "max. " << *cp.max() << ", ";
-                            }
-                            ss << "Press Enter to send";
-                            break;
-                        case webcface::ViewComponentType::select_input:
-                            ss << "Press Enter of Space to select";
-                            break;
-                        default:
-                            break;
-                        }
-                        *help = ss.str();
-                    }
-                }
-                break;
-            }
-            }
-        }
-        elements.push_back(RenderCol(elements_col, elements.size()));
-        return ftxui::window(ftxui::hbox({
-                                 ftxui::text(view.member().name()),
-                                 ftxui::text(":"),
-                                 ftxui::text(view.name()),
-                             }),
-                             ftxui::vbox(elements));
-    }
+    ftxui::Element Render() override;
 };
-inline ftxui::Component viewComponent(const webcface::View &view,
-                                      std::shared_ptr<std::string> help,
-                                      std::shared_ptr<ftxui::Element> result,
-                                      bool light) {
-    return std::make_shared<ViewUIContainer>(view, help, result, light);
-}
-inline void
-addViewComponent(ftxui::ScreenInteractive &screen, ftxui::Component &container,
-                 const webcface::View &view, std::shared_ptr<std::string> help,
-                 std::shared_ptr<ftxui::Element> result, bool light) {
-    auto handle = std::make_shared<webcface::View::EventHandle>();
-    *handle = view.prependListener([=, &screen, &container] {
-        screen.Post([=, &container] {
-            container->Add(viewComponent(view, help, result, light));
-        });
-        view.removeListener(*handle);
-    });
-}
+
+ftxui::Component viewComponent(const webcface::View &view,
+                               std::shared_ptr<std::string> help,
+                               std::shared_ptr<ftxui::Element> result,
+                               bool light);
+void addViewComponent(ftxui::ScreenInteractive &screen,
+                      ftxui::Component &container, const webcface::View &view,
+                      std::shared_ptr<std::string> help,
+                      std::shared_ptr<ftxui::Element> result, bool light);
